# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Property Portal API",
        "description": "API documentation for the Property Portal backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/properties/{slug}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get reviews for a property",
                "operationId": "db914b4ab9cb71f433152d3896bf3a13",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reviews for the property",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/reviews": {
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Submit a review for a booking",
                "operationId": "faeca7cb1858950e575451fd5d97670e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_id",
                                    "ratings",
                                    "comment"
                                ],
                                "properties": {
                                    "booking_id": {
                                        "type": "string",
                                        "example": "BK-0982"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Amazing stay! Loved the view."
                                    },
                                    "ratings": {
                                        "properties": {
                                            "amenities": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "cleanliness": {
                                                "type": "integer",
                                                "example": 4
                                            },
                                            "communication": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "location": {
                                                "type": "integer",
                                                "example": 4
                                            },
                                            "value": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review submitted"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/customer/reviews/{id}": {
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Update your review",
                "operationId": "e0d96c05f2be0a1aa7ea6c424efe6e1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_id",
                                    "ratings",
                                    "comment"
                                ],
                                "properties": {
                                    "booking_id": {
                                        "type": "string",
                                        "example": "BK-0982"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Amazing stay! Loved the view."
                                    },
                                    "ratings": {
                                        "properties": {
                                            "amenities": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "cleanliness": {
                                                "type": "integer",
                                                "example": 4
                                            },
                                            "communication": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "location": {
                                                "type": "integer",
                                                "example": 4
                                            },
                                            "value": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete your review",
                "operationId": "c923cb491d7a99a9b8da7400de7e48a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}